requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
psycopg2-binary
python-jose[cryptography]  # JWT
passlib[bcrypt]            # password hashing
python-dotenv
pydantic[email]
qrcode[pil]
redis
celery

